+++ src/converter.ts
@@ -1,9 +1,14 @@
+import { URL } from 'url'
 import { MarpOptions } from '@marp-team/marp-core'
 import { Marpit, MarpitOptions } from '@marp-team/marpit'
 import chalk from 'chalk'
-import * as chromeFinder from 'chrome-launcher/dist/chrome-finder'
 import puppeteer from 'puppeteer-core'
-import { URL } from 'url'
+import {
+  generatePuppeteerDataDirPath,
+  generatePuppeteerLaunchArgs,
+  isWSL,
+  resolveWSLPath,
+} from './utils/puppeteer'
 import { silence, warn } from './cli'
 import { Engine } from './engine'
 import metaPlugin from './engine/meta-plugin'
@@ -48,7 +53,6 @@ export interface ConverterOption {
   pages?: boolean | number[]
   preview?: boolean
   jpegQuality?: number
-  readyScript?: string
   server?: boolean
   template: string
   templateOption?: TemplateOption
@@ -91,7 +95,7 @@ export class Converter {
   }
 
   async convert(markdown: string, file?: File): Promise<TemplateResult> {
-    const { lang, readyScript, globalDirectives, type } = this.options
+    const { lang, globalDirectives, type } = this.options
     const isFile = file && file.type === FileType.File
 
     let additionals = ''
@@ -107,7 +111,6 @@ export class Converter {
     return await this.template({
       ...(this.options.templateOption || {}),
       lang,
-      readyScript,
       base:
         isFile && type !== ConvertType.html
           ? file!.absoluteFileScheme
@@ -345,9 +348,13 @@ export class Converter {
       return baseFile.saveTmpFile('.html')
     })()
 
-    const uri = tmpFile
-      ? `file://${tmpFile.path}`
-      : `data:text/html;base64,${baseFile.buffer!.toString('base64')}`
+    const uri = await (async () => {
+      if (tmpFile) {
+        if (isWSL()) return `file:${await resolveWSLPath(tmpFile.path)}`
+        return `file://${tmpFile.path}`
+      }
+      return `data:text/html;base64,${baseFile.buffer!.toString('base64')}`
+    })()
 
     try {
       const browser = await Converter.runBrowser()
@@ -396,25 +403,10 @@ export class Converter {
 
   private static async runBrowser() {
     if (!Converter.browser) {
-      const args: string[] = []
-      if (process.env.IS_DOCKER) args.push('--no-sandbox')
-
-      // Workaround for Chrome 73 in docker and unit testing with CircleCI
-      // https://github.com/GoogleChrome/puppeteer/issues/3774
-      if (process.env.IS_DOCKER || process.env.CI)
-        args.push('--disable-features=VizDisplayCompositor')
-
-      const finder: () => string[] = (() => {
-        if (process.env.IS_DOCKER) return () => ['/usr/bin/chromium-browser']
-        if (require('is-wsl')) return chromeFinder.wsl
-        return chromeFinder[process.platform]
-      })()
-
       Converter.browser = await puppeteer.launch({
-        args,
-        executablePath: finder ? finder()[0] : undefined,
+        ...(await generatePuppeteerLaunchArgs()),
+        userDataDir: await generatePuppeteerDataDirPath('marp-cli-conversion'),
       })
-
       Converter.browser.once('disconnected', () => {
         Converter.browser = undefined
       })